+++ fix-lint-backend.sh
@@ -16,5 +16,4 @@ npm run fix
 ###################
 ###### Shell ######
 ###################
-#shfmt -s -w -i 4 ./*.sh ./**/*.sh
 shellharden --replace ./*.sh ./**/*.sh